lib: introduce ConnectPolicy
ConnectPolicy is to be used on a client side of a connection to allow
user to decide on the method to perform handshake. To either use
anonymous handshake or login/pass combination without the need of
separate api for that. Also this approach will support sessions that
will be implemented later without the need to change client side api.
Implement SimpleConnectPolicy that always performs anonymous handshake.